requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coverage==7.12.0
cryptography==46.0.3
cssselect2==0.8.0
cv==1.0.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
fastapi==0.119.1
freetype-py==2.5.1
git-filter-repo==2.47.0
h11==0.16.0
html5lib==1.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
oscrypto==1.3.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.3.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
svglib==1.6.0
tinycss2==1.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.38.0
waitress==3.0.2
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
